Make Ann a record and use it throughout
By making it a record and not relying on the non-record syntax, we can
easily reorder or add new fields to it in the future without breaking
any existing patterns.

This commit also uses the record update syntax where before an entirely
new Ann was constructed, saving a bunch of variables (and implicitly
passing through any extra fields added to Ann in the future).