Always preserve expanded attribute sets
This changes the formatting to take into account whether attribute sets
are expanded onto multiple lines already, and if so, preserves that.

This is desirable because if a user writes an attribute set over
multiple lines, there's most likely an intention to add more attributes,
which would become harder when formatted.

An example of this is module system options, which usually start out
with just a type (or even no attributes), but usually get more
attributes later. This would've been contracted onto a single line
before this change:

    options.foo = lib.mkOption {
      type = types.str;
    };

This change conforms to the standard due to this line:

> The formatter may take the input formatting into account in some cases in order to preserve multi-line syntax elements (which would otherwise have been contracted by the rules).